_IViewComponentResult
To divide the product page into two sections—menu and product—I created a "CategoriesMenuViewComponent.cs" within the Components directory. This component retrieves all categories and displays them as an enumerable in the Default.cshtml page using a foreach loop.

In the product page, I used the `@await Component.InvokeAsync("CategoriesMenu")` method. This method automatically executes the necessary methods due to the convention structure, thus achieving the desired layout for the project.